Support directories in `-doc-external-doc`
It is documented as accepting a “classpath_entry_path” for the keys but
this only worked for JARs and not for individual class files. When
checking for external-doc mappings for a Symbol, we now find the root
directory relative to a class file instead of using the full class file
path.

The corresponding tests for SI-191 and SI8557 are also fixed to support
individual class files instead of JARs in partest. This is required for
the sbt build which runs partest on “quick” instead of “pack”.