wip: Fix scaladoc comments of Map classes
Due to the way comments are inherited in scaladocs, there is no way of
overriding a comment without overriding the corresponding method with
an explicit `super` call in some cases.

This affects flatten in GenMapLike/GenMap, a methods which should not
exist in the first place because you can never call it. The effect in
the various Map/MapLike classes is much worse. They do not inherit the
GenMapLike comments and since the implementation happens elsewhere, they
cannot override the comments, either.

++ in MapLike / Map: Simply adding the required declaration makes calls
to ++ in other places ambiguous even though it doesn’t add a new
overload (the method is already overloaded).

I did not bother changing documentation of zip/zipAll/zipWithIndex.
They still show the resulting Map types with an extra set of parentheses
in the use cases but these use cases are generally useless for maps
anyway.