An attempt at simplifying WithFilter
`collection.WithFiler` should not exist in the first place. It
pretends that there is a “generic” version of `WithFilter` but that’s
not true. There is one version per collection kind. Non-Iterable
types use completely unrelated WithFilter classes.

WithFilter should exist in different versions for different collection
kinds, stacked on top of each other in the same way as Ops traits.

The remaining problem is the `withFilter` method. Its return type has
to be the appropriate `WithFilter` type but if `CC` types of different
kinds are refined, this will never work automatically. For example,
`mutable.Map` has an `IterableCC` of `mutable.Iterable` but it can
only refine the `MapCC` in `MapOps.WithFilter`.

Threading all parent `CC` types through `WithFilter` wouldn’t help
because they are not available at the points where `withFilter` is
defined. Refinements are done by mixing in different Ops traits but
that only works when each method requires only *one* of the `CC`
types. `WithFilter` requires *all* of them but due to linearization
only one can be preserved.

It looks like we will have to wait for intersection types to make
this simpler design work.